pyproject.toml
[project]
name = "ai-service"
version = "0.1.0"
description = "Python版图片处理服务"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.11",
    "uvicorn==0.34.0",
    "openai",
    "pydantic==2.6.1",
    "pydantic-settings",
    "python-multipart==0.0.20",
    "redis==5.0.1",
    "minio==7.2.15",
    "numpy<2",
    "pillow==11.1.0",
    "python-dotenv==1.0.1",
    "setuptools==76.1.0",
    "torch==2.5.1",
    "torchvision==0.20.1",
    "opencv-python==4.11.0.86",
    "timm==1.0.15",
    "scipy==1.15.2",
    "scikit-image==0.25.2",
    "kornia==0.8.0",
    "einops==0.8.1",
    "transformers==4.49.0",
    "huggingface-hub==0.29.3",
    "accelerate==1.5.2",
]

[project.scripts]
start = "uvicorn main:app --host 0.0.0.0 --port 8000 --reload"
